+++ src/HDInsight/HDInsight/ChangeLog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Supported creating cluster with encryption at host feature.
 
 ## Version 3.4.0
 * Supported creating cluster with encryption in transit feature.